requirements.txt
django
djangorestframework
pytils
